# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "todolist API",
        "title": "todolist API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/myungjin-likes-tuna/checkareer-core-go/issues",
            "email": "nexters@kakao.com"
        },
        "version": "21.0.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/skills": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "skills get",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "skills get",
                "operationId": "skills-get",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 25,
                        "example": 25,
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skills.ManyItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "skills.ManyItemResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/skills.Node"
                    }
                }
            }
        },
        "skills.Node": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}